move tests to test_sequential.py
